Se mejoro la migracion de la base de datos para permitir valores mas grades en la cedula y telefono, adicional a eso se modificio el metodo create para evitar un error con el terminos y condiciones y por ultimo se cambio la ruta para mostrar mensaje de exito

+++ database/migrations/2022_08_16_135834_create_participantes_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
             $table->id('cod_participantes');
             $table->string('nombre', 50);
             $table->string('apellido', 50);
-            $table->integer('cedula')->unique();
-            $table->integer('telefono')->unique();
+            $table->bigInteger('cedula')->unique();
+            $table->bigInteger('telefono')->unique();
             $table->bigInteger('cod_departamento')->unsigned();
             $table->bigInteger('cod_ciudad')->unsigned();
             $table->string('correo', 80)->unique();
